A lot of rework was made to the project´s logic.
Reports were added to Tienda.java, methods to add Pc components were added and the one regarding RAM was fully programmed, also the passive atribute "marca", since it was repeated in each Component class, was instead placed at the Superclass and the rest just inherit the atribute. The constructors also changed a bit due to this.