Update MenuTienda.java
JtextFIelds with limits were added, the rest of the panels that concern about entering info about a pc component were added, only the RAM one is fully functional, the rest is just a placeholder, the workers and components list now doesnt appears unless at least one of them were entered in the program. A new menu was added to display reports, and the oldest one was renamed to lists.